feat(v0.3.0): Real AI noise cancellation, dependency overhaul, release infrastructure
Wire real RNNoise-based noise cancellation into the DSP pipeline via
nnnoiseless (pure-Rust port with pre-trained weights). At 48kHz the
pipeline now runs neural network inference on CPU with automatic
fallback to spectral denoising at other sample rates.

Core changes:
- Add nnnoiseless crate for real AI noise reduction (480-sample frames, 10ms latency)
- NnnoiseDenoiser adapter with ring buffer for arbitrary pipeline buffer sizes
- DenoiserBackend enum (Nnnoiseless/Spectral) with runtime switching
- Auto-select nnnoiseless at 48kHz, spectral fallback at 96kHz+

Dependency overhaul:
- cudarc 0.11 → 0.19 (CudaDevice→CudaContext, stream-based API)
- cpal 0.15 → 0.17, alsa 0.9 → 0.11, jack 0.11 → 0.13
- PipeWire 0.8 → 0.9, Edition 2024, Rust 1.85+
- Remove dead vulkan-compute and opencl feature flags
- Fix license metadata to MIT (was incorrectly MIT OR Apache-2.0)

Release infrastructure:
- Consolidate 3 PKGBUILDs into single release/arch/PKGBUILD
- Add release/install-system.sh universal installer
- Add GitHub issue templates (bug report, feature request, hardware compat)
- Reorganize docs/ into backends/, gpu/, integration/, development/
- Honest beta README messaging (working vs in-development features)

241 tests passing, 0 failures.

+++ src/jack_module.rs
@@ -150,7 +150,7 @@ impl JackModule {
 
         // Verify sample rate compatibility
         let jack_sample_rate = client.sample_rate();
-        if jack_sample_rate != self.config.audio.sample_rate as usize {
+        if jack_sample_rate != self.config.audio.sample_rate {
             warn!(
                 "Sample rate mismatch: JACK={}Hz, Config={}Hz",
                 jack_sample_rate, self.config.audio.sample_rate
@@ -328,7 +328,7 @@ impl Drop for JackModule {
 #[allow(dead_code)] // Public API for JACK introspection
 pub struct JackInfo {
     pub client_name: String,
-    pub sample_rate: usize,
+    pub sample_rate: u32,
     pub buffer_size: u32,
     pub cpu_load: f32,
     pub is_realtime: bool,
